[Amp/Fix] Properly save AMPs as ushorts instead of bytes

+++ Source/NexusForever.WorldServer/Game/Spell/ActionSetAmp.cs
@@ -44,7 +44,7 @@ namespace NexusForever.WorldServer.Game.Spell
             {
                 Id        = actionSet.Owner,
                 SpecIndex = actionSet.Index,
-                AmpId     = (byte)Entry.Id
+                AmpId     = (ushort)Entry.Id
             };
 
             if ((saveMask & AmpSaveMask.Create) != 0)
